fix warnings

+++ src/selection.rs
@@ -69,7 +69,7 @@ fn mouse_button_input(
         let world_pos = screen_to_world(
             &transform,
             cursor_position,
-            Vec2::new(window.width() as f32, window.height() as f32),
+            Vec2::new(window.width(), window.height()),
         );
 
         selection_rect.first_x = Some(world_pos.x);
@@ -90,7 +90,7 @@ fn mouse_button_input(
         let world_pos = screen_to_world(
             &transform,
             cursor_position,
-            Vec2::new(window.width() as f32, window.height() as f32),
+            Vec2::new(window.width(), window.height()),
         );
 
         let (x1, x2) = (some_or_return!(selection_rect.first_x), world_pos.x);
@@ -130,7 +130,7 @@ fn mouse_button_input(
         let world_pos = screen_to_world(
             &transform,
             cursor_position,
-            Vec2::new(window.width() as f32, window.height() as f32),
+            Vec2::new(window.width(), window.height()),
         );
 
         selection_rect.second_x = Some(world_pos.x);
